Correct type for quantity consumed

+++ Agent-Based Simulation/header/Person.h
@@ -63,5 +63,5 @@ class Person : public Agent {
     void log_inventory();
     void log_account();
     void log_health_status();
-    void log_consumption(const Product * product, const double quantity);
+    void log_consumption(const Product * product, const int quantity);
 };
